* commit '8bbfc9c8e1eead98fefe2ec256daf8b8d1b26dd0':
  Improve logging

+++ src/core/config/AdaVisionConfig.h
@@ -124,7 +124,7 @@ public:
 
         std::string zmqAddress() const;
 
-        int mockHeading() const;
+        double mockHeading() const;
 
     private:
         const dlib::config_reader &_currentDataConfig;
